Remove the necessity of global variables
In Kainda Internal, whenever there is a need to use the Models variable, we recalculate it. It's not the most optimal way but is easy to understand and doesnt clutter the generated code.
In the generated project Models are stored inside ModelsService class, that can be imported anytime.

Same problem solved with sequelize and mongoose using the DbService class and passing the values from the app to the kainda internal when needed, like in transactions