Transactions commit or rollback only if active
Commit and rollback functions now check if the transaction is active before trying to execute the operation, so no exception is thrown if it is used incorrectly.
Added option to throw an exception if the transaction state is not expected. i.e calling commit on a transaction already commited or rolled back